--- native/opencv-wrapper.cc
@@ -1,74 +1,0 @@
-#include <opencv2/core.hpp>
-#include <opencv2/features2d.hpp>
-#include <opencv2/highgui.hpp>
-#include <opencv2/imgproc.hpp>
-#include <opencv2/objdetect.hpp>
-#include <opencv2/text/ocr.hpp>
-#include <opencv2/video/tracking.hpp>
-#include <opencv2/xfeatures2d.hpp>
-
-#include "opencv-wrapper.h"
-#include "utils.h"
-
-EXTERN_C_BEGIN
-
-// =============================================================================
-//   Core
-// =============================================================================
-
-void cv_vec_drop(CVec<void>* vec, unsigned int depth) {
-    if (vec->array != nullptr) {
-        if (depth > 1) {
-            auto nestedVec = (CVec<void>*) vec->array;
-            for (size_t i = 0; i < vec->size; ++i) {
-                cv_vec_drop(&nestedVec[i], depth - 1);
-            }
-        }
-        free(vec->array);
-        vec->array = nullptr;
-        vec->size = 0;
-    }
-}
-
-void c_drop(void* value) {
-    free(value);
-    value = nullptr;
-}
-
-// =============================================================================
-//  core array
-// =============================================================================
-
-// =============================================================================
-//  Imgproc
-// =============================================================================
-
-// =============================================================================
-//  Imgcodecs
-// =============================================================================
-
-// =============================================================================
-//   Highgui: high-level GUI
-// =============================================================================
-
-// =============================================================================
-//   VideoCapture
-// =============================================================================
-
-// =============================================================================
-//   VideoWriter
-// =============================================================================
-
-// =============================================================================
-//   CascadeClassifier
-// =============================================================================
-
-// =============================================================================
-//  Object Tracking
-// =============================================================================
-
-// =============================================================================
-//   MSER
-// =============================================================================
-
-EXTERN_C_END
