--- native/opencv-wrapper.h
@@ -1,62 +1,0 @@
-#ifndef OPENCV_WRAPPER_H_
-#define OPENCV_WRAPPER_H_
-
-#include <functional>
-#include <opencv2/core.hpp>
-#include <opencv2/features2d.hpp>
-#include <opencv2/highgui.hpp>
-#include <opencv2/imgproc.hpp>
-#include <opencv2/objdetect.hpp>
-#include <opencv2/text/ocr.hpp>
-#include <opencv2/video/tracking.hpp>
-#include <opencv2/xfeatures2d.hpp>
-#include <stddef.h>
-#include <stdint.h>
-
-#include "common.h"
-
-#define EXTERN_C_BEGIN extern "C" {
-#define EXTERN_C_END }
-
-EXTERN_C_BEGIN
-
-// The caller owns the returned data cv::Mat
-
-void cv_vec_drop(CVec<void>* vec, unsigned int depth);
-void c_drop(void* value);
-
-// =============================================================================
-//  core array
-// =============================================================================
-
-// =============================================================================
-//  Imgproc
-// =============================================================================
-
-// =============================================================================
-//  Imgcodecs
-// =============================================================================
-
-// =============================================================================
-//   Highgui: high-level GUI
-// =============================================================================
-
-// =============================================================================
-//
-// =============================================================================
-
-// =============================================================================
-//   CascadeClassifier
-// =============================================================================
-
-// =============================================================================
-//   VideoTrack
-// =============================================================================
-
-// =============================================================================
-//   Text
-// =============================================================================
-
-EXTERN_C_END
-
-#endif  // OPENCV_WRAPPER_H_
